+++ projects/rocprofiler-compute/tests/test_pc_sampling.py
@@ -55,6 +55,8 @@ def test_pc_sampling_host_trap(binary_handler_profile_rocprof_compute):
     common.skip_unsupported_pc_sampling_soc()
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "--pc-sampling-method",
@@ -87,6 +89,8 @@ def test_pc_sampling_stochastic(binary_handler_profile_rocprof_compute):
     common.skip_unsupported_pc_sampling_soc(is_stochastic=True)
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "--pc-sampling-method",
@@ -134,6 +138,8 @@ def test_multi_rank_pc_sampling_only(
     workload_dir = common.get_output_dir()
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "--pc-sampling-method",
@@ -173,6 +179,8 @@ def test_multi_rank_warning_pc_sampling_with_counters(
     workload_dir = common.get_output_dir()
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "2",
@@ -213,6 +221,8 @@ def test_pc_sampling_profile_then_analyze(
     common.skip_unsupported_pc_sampling_soc()
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "--pc-sampling-method",
@@ -240,6 +250,8 @@ def test_pc_sampling_profile_then_analyze(
             "analyze",
             "--path",
             workload_dir,
+            "--experimental",
+            "--pc-sampling",
             "--block",
             "21",
         ],
@@ -271,6 +283,8 @@ def test_pc_sampling_profile_then_analyze(
             "analyze",
             "--path",
             workload_dir,
+            "--experimental",
+            "--pc-sampling",
             "--block",
             "21",
             "--kernel",
@@ -295,6 +309,8 @@ def test_pc_sampling_with_sol_block(binary_handler_profile_rocprof_compute):
     common.skip_unsupported_pc_sampling_soc()
 
     options = [
+        "--experimental",
+        "--pc-sampling",
         "--block",
         "21",
         "2",